refactored final strings into designated files
