+++ src/main/java/apple/commands/ScrollHelpCommand.java
@@ -14,7 +14,7 @@ public class ScrollHelpCommand implements CommandExecutor {
         this.plugin = plugin;
         PluginCommand command = plugin.getCommand("scroll_help");
         if (command == null) {
-            System.out.println("ScrollsTp could not get the scrollDuplicate command");
+            System.err.println("ScrollsTp could not get the scrollDuplicate command");
             return;
         }
         command.setExecutor(this);
